Updating Multiscale Entropy Function
Trying to replicate methods from another study I realized that there may be issues with this and the pyentrp implementations of MSE. emb_dim is the length of the vectors to compare. However, in this implementation it is used as the scale_factor as defined in the literature. It appears that the true place that emb_dim should be placed (in the sample entropy calculation) is hard coded as one. I have added a new input called max_scale_factor (tau in the costa paper referenced) - this will ensure that the for loop iterates over all scales between 1 and the inputted max_scale_factor. I placed the emb_dim in the locations it should be in the sample entropy function (replacing the hard coded 1). Last, I think the output of this function is incorrect. From my brief viewing of the literature it seems there are two general outputs that may be used. 1) the sample entropy for each of the individual scale_factors (1:max_scale_factor defined above), or 2) the area under a curve created by these individual scale factors. I have added a variable that allows the user to output either a single value, or the data from all of the scale_factors. I have not changed the output of the individual value that is outputted, but have included a comment with that I think is going on and a one line implementation that I beleive will produce the area output described.